pyproject.toml
[tool.poetry]
name = "pong"
version = "0.1.0"
description = "PyGame implementation of the classic game Pong"
authors = ["Mike Watters <valholla@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.4.1"

[tool.poetry.group.Dev.dependencies]
pylint = "3.*"
pytest = "8.*"
isort = "5.*"
black = "24.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120
skip_glob = ["tmp/*"]
src_paths = ["game", "**"]

[tool.black]
line-length = 120
preview = true
#include = '/(game|pong)/'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | __pycache__
    | \.tox
    | \.venv
    | venv
    | \.vscode
    | \.env
    | env
    | _build
    | build
    | dist
    | tmp
    | tests/fixtures
    | game/fonts
  )/
  | example
)
'''

[tool.pylint.main]
ignore-paths = [".eggs",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
    ".tox",
    ".venv",
    "venv",
    ".env",
    "env",
    "_build",
    "build",
    "dist",
    "tmp"
]

[tool.pylint.messages_control]
max-line-length = 120
good-names = [
"k", "v", "e", "f", "_"
]
disable = [
  "anomalous-backslash-in-string",
  "logging-fstring-interpolation",
  "missing-docstring",
  "missing-class-docstring",
  "missing-function-docstring",
  "missing-module-docstring",
  "no-else-return",
  "no-member",
  "no-value-for-parameter",
  "protected-access",
  "too-few-public-methods",
  "too-many-arguments",
  "too-many-instance-attributes",
]

[tool.pylint.design]
# limiting the number of returns might discourage
# the use of guard clauses. So we increase the
# allowed number of returns from 6 to 8
max-returns = 8

